Refactor database module and create new utility file
Refactored database module by moving various utility functions from the `query.py` file to a newly created `connection_processing.py` file. Updated all references to the moved functions across multiple modules and tests to ensure consistency. This cleanup aims to improve maintainability and clarity of the codebase.
