+++ modules/utils/database/crossref.py
@@ -64,7 +64,7 @@ def process_publication_info(doi: str, result: Dict):
 
     from modules.building_block import Article
 
-    article: Article = ArticleFactory.get_object(doi)
+    article: Article = ArticleFactory.get_base_object(doi)
     # get publication list
     tmp = process_link(result['link'], 'content-type', 'application/xml')
     if tmp is None or valid_link(tmp) is False:
@@ -93,7 +93,7 @@ def process_link(links: List[Dict], key: str, value: str) -> Union[str, None]:
 
 
 def valid_link(url: str) -> bool:
-    from modules.utils.database.query import connect_url
+    from modules.utils.database.connection_processing import connect_url
 
     from requests import HTTPError
 
@@ -109,7 +109,7 @@ if __name__ == '__main__':
     # get_publication_info('10.5555/515151')
     # get_publication_info("10.1101/104778")
     factory = ArticleFactory()
-    print(factory.create_object(identifier="10.1038/s41598-017-04402-4"))
+    print(factory.create_base_object(identifier="10.1038/s41598-017-04402-4"))
     get_publication_info("10.1038/s41598-017-04402-4")
     # get_publication_info("10.1109/MM.2019.2910009")
     # valid_link('https://www.nature.com/articles/s41598-017-04402-4.pdf')
